Adapt to use a form type
