+++ src/Action/RequestAction.php
@@ -16,8 +16,12 @@ namespace Sonata\UserBundle\Action;
 use Sonata\AdminBundle\Admin\Pool;
 use Sonata\AdminBundle\Templating\TemplateRegistryInterface;
 use Sonata\UserBundle\Form\Type\ResetPasswordRequestFormType;
+use Sonata\UserBundle\Mailer\MailerInterface;
+use Sonata\UserBundle\Model\UserManagerInterface;
+use Sonata\UserBundle\Util\TokenGeneratorInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
@@ -37,13 +41,25 @@ final class RequestAction
 
     private FormFactoryInterface $formFactory;
 
+    private UserManagerInterface $userManager;
+
+    private MailerInterface $mailer;
+
+    private TokenGeneratorInterface $tokenGenerator;
+
+    private int $retryTtl;
+
     public function __construct(
         Environment $twig,
         UrlGeneratorInterface $urlGenerator,
         AuthorizationCheckerInterface $authorizationChecker,
         Pool $adminPool,
         TemplateRegistryInterface $templateRegistry,
-        FormFactoryInterface $formFactory
+        FormFactoryInterface $formFactory,
+        UserManagerInterface $userManager,
+        MailerInterface $mailer,
+        TokenGeneratorInterface $tokenGenerator,
+        int $retryTtl
     ) {
         $this->twig = $twig;
         $this->urlGenerator = $urlGenerator;
@@ -51,15 +67,39 @@ final class RequestAction
         $this->adminPool = $adminPool;
         $this->templateRegistry = $templateRegistry;
         $this->formFactory = $formFactory;
+        $this->userManager = $userManager;
+        $this->mailer = $mailer;
+        $this->tokenGenerator = $tokenGenerator;
+        $this->retryTtl = $retryTtl;
     }
 
-    public function __invoke(): Response
+    public function __invoke(Request $request): Response
     {
         if ($this->authorizationChecker->isGranted('IS_AUTHENTICATED_FULLY')) {
             return new RedirectResponse($this->urlGenerator->generate('sonata_admin_dashboard'));
         }
 
         $form = $this->formFactory->create(ResetPasswordRequestFormType::class);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $username = $form->get('username')->getData();
+            $user = $this->userManager->findUserByUsernameOrEmail($username);
+
+            if (null !== $user && $user->isEnabled() && !$user->isPasswordRequestNonExpired($this->retryTtl) && $user->isAccountNonLocked()) {
+                if (null === $user->getConfirmationToken()) {
+                    $user->setConfirmationToken($this->tokenGenerator->generateToken());
+                }
+
+                $this->mailer->sendResettingEmailMessage($user);
+                $user->setPasswordRequestedAt(new \DateTime());
+                $this->userManager->save($user);
+            }
+
+            return new RedirectResponse($this->urlGenerator->generate('sonata_user_admin_resetting_check_email', [
+                'username' => $username,
+            ]));
+        }
 
         return new Response($this->twig->render('@SonataUser/Admin/Security/Resetting/request.html.twig', [
             'base_template' => $this->templateRegistry->getTemplate('layout'),
