+++ src/Resources/config/routing/admin_resetting.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <routes xmlns="http://symfony.com/schema/routing" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/routing http://symfony.com/schema/routing/routing-1.0.xsd">
-    <route id="sonata_user_admin_resetting_request" path="/request" methods="GET" controller="sonata.user.action.request"/>
-    <route id="sonata_user_admin_resetting_send_email" path="/send-email" methods="POST" controller="sonata.user.action.send_email"/>
+    <route id="sonata_user_admin_resetting_request" path="/request" methods="GET POST" controller="sonata.user.action.request"/>
     <route id="sonata_user_admin_resetting_check_email" path="/check-email" methods="GET" controller="sonata.user.action.check_email"/>
     <route id="sonata_user_admin_resetting_reset" path="/reset/{token}" methods="GET POST" controller="sonata.user.action.reset"/>
 </routes>
